% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Forecast and evaluate the models with RMSE}
\usage{
calculate_rmse(fbl, test, target)
}
\arguments{
\item{fbl}{A \code{fable} with predictions.}

\item{test}{A \code{tsibble} with test data.}

\item{target}{A \code{character} indicating the target variable.}
}
\value{
A \code{tibble} with RMSE for each model.
}
\description{
Forecast and evaluate the models with RMSE
}
