% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{converter}
\alias{converter}
\title{Convert product size to volume (ml) and mass (oz)}
\usage{
converter(x)
}
\arguments{
\item{x}{A \code{character} containing the product size.}
}
\value{
A \code{number} of the converted product size.
}
\description{
Convert product size to volume (ml) and mass (oz)
}
\keyword{internal}
