% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{data_loading}
\alias{data_loading}
\title{Data Loading and Preparation}
\usage{
data_loading(raw_data_path, data_path)
}
\arguments{
\item{raw_data_path}{Optional path to the raw data file location.}

\item{data_path}{Directory where the output files will be saved.}
}
\value{
A list containing the processed data frames: store, product, and transaction.
}
\description{
If raw data path is provided, apply the processing steps. Otherwise, use the default datasets.
Save them to the output directory.
}
