% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{dh_forecasting}
\alias{dh_forecasting}
\title{Forecasting}
\usage{
dh_forecasting(
  steps = 2,
  raw_data_path = NULL,
  output_dir = getwd(),
  batch_size = 8
)
}
\arguments{
\item{steps}{Number of steps to forecast ahead. Defaults to 2.}

\item{raw_data_path}{Optional path to the raw data file location.}

\item{output_dir}{Directory where the output files will be saved. Defaults to the current working directory.}

\item{batch_size}{The size of the batch for forecasting. Defaults to 8.}
}
\description{
Models:
\itemize{
\item \strong{ARIMA}:
\itemize{
\item \code{arima_def = fable::ARIMA(sqrt(units))}
\item \code{arima_lagged = fable::ARIMA(sqrt(units) ~ feature + display + tpr_only + dplyr::lag(feature) + dplyr::lag(display) + dplyr::lag(tpr_only))}
(Lagged ARIMA model is preferred over the default ARIMA model (if it isn't null))
}
\item \strong{STL + ETS model}:
After interpolating missing values using the selected ARIMA model,
an STL decomposition followed by ETS modeling is performed.
}
}
\seealso{
\code{\link[fable:ARIMA]{fable::ARIMA}} \code{\link[feasts:STL]{feasts::STL}} \code{\link[fable:ETS]{fable::ETS}}
}
