% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{dh_validation}
\alias{dh_validation}
\title{Validation}
\usage{
dh_validation(
  raw_data_path = NULL,
  output_dir = getwd(),
  test_size = 0.2,
  batch_size = 8
)
}
\arguments{
\item{raw_data_path}{Optional path to the raw data file location.}

\item{output_dir}{Directory where the output files will be saved. Defaults to the current working directory.}

\item{test_size}{Proportion of the dataset to be used for testing.}

\item{batch_size}{The size of the batch for forecasting. Defaults to 8.}
}
\description{
Validation
}
