% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_ts}
\alias{filter_ts}
\title{Filter TS}
\usage{
filter_ts(transaction)
}
\arguments{
\item{transaction}{A \code{tsibble} containing the transaction data.}
}
\value{
A \code{tsibble} with the filtered time series.
}
\description{
Conserve those TS with more than 12 observations (3 months at least).
}
