% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_forecast}
\alias{get_forecast}
\title{Forecasting function}
\usage{
get_forecast(mbl, test, bs)
}
\arguments{
\item{mbl}{A \code{mable} with fitted models.}

\item{test}{A \code{tsibble} with test data.}

\item{bs}{An \code{integer} with batch size.}
}
\value{
A \code{fable} with forecasts for the test data.
}
\description{
Forecasting function
}
