% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hilo_bound}
\alias{hilo_bound}
\title{Helper function to calculate lower and upper bounds}
\usage{
hilo_bound(fbl, target, level, pos)
}
\arguments{
\item{fbl}{A \code{fable} object with forecasts.}

\item{target}{A \code{character} indicating the target variable.}

\item{level}{A \code{numeric} with the confidence level.}

\item{pos}{A \code{character} with the position of the bound (\code{lower} or \code{upper}).}
}
\value{
A \code{numeric} with the bound or NA if there is an error.
}
\description{
Helper function to calculate lower and upper bounds
}
