% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_ts}
\alias{impute_ts}
\title{Imputation}
\usage{
impute_ts(sampled_transaction)
}
\arguments{
\item{sampled_transaction}{A \code{tsibble} with sampled time series.}
}
\value{
A \code{tsibble} with the imputed time series.
}
\description{
Nulls imputation with ARIMA because some models require non-nulls.
Also, I impute feature, display and tpr_only with 0.
}
