% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Load raw data}
\usage{
load_raw_data(raw_data_path)
}
\arguments{
\item{raw_data_path}{A \code{character} with the path to the raw data Excel file.}
}
\value{
A \code{list} of dataframes for store, product and transaction datasets.
}
\description{
Load raw data from Excel and clean their column names.
It has to be in the same format as the original data.
}
