% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_transaction}
\alias{missing_transaction}
\title{Missing transaction}
\usage{
missing_transaction(tsb)
}
\arguments{
\item{tsb}{A \code{tsibble} containing the transaction data.}
}
\value{
A \code{tsibble} with the number of missing transaction by store and product.
}
\description{
Aggregate missing values accross key groups.
}
