% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(data)
}
\arguments{
\item{data}{A \code{list} of dataframes for store, product and transaction.}
}
\value{
A \code{list} of processed dataframes.
}
\description{
Prepare data to comsumption. Steps for each dataset:
\itemize{
\item Store -> Unify repeated store with distinct \code{seg_value_name}.
\item Product -> Convert product size to volume (ml) and mass (oz).
\item Transaction -> Parse to \code{tsibble} format.
}
}
