% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesforecastr-package.R
\docType{package}
\name{salesforecastr-package}
\alias{salesforecastr}
\alias{salesforecastr-package}
\title{salesforecastr: Sales Forecasting with Time Series Models}
\description{
A package for sales forecasting using time series models. It provides tools for data preparation, model fitting, and evaluation, with a focus on ease of use and flexibility.
}
\author{
\strong{Maintainer}: Alejandro Vargas \email{alejandrovargasauger@gmail.com}

}
\keyword{internal}
