% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sample_for_debugging}
\alias{sample_for_debugging}
\title{Sample TS for debugging}
\usage{
sample_for_debugging(ts, n = 10)
}
\arguments{
\item{ts}{A \code{tsibble} containing the time series.}

\item{n}{Number of samples to select.}
}
\description{
Sample TS for debugging
}
