% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_ts}
\alias{sample_ts}
\title{Sample TS}
\usage{
sample_ts(transaction, product)
}
\arguments{
\item{transaction}{A \code{tsibble} with the number of missing transaction by store and product.}

\item{product}{A \code{tibble} containing the product data.}
}
\value{
A \code{tsibble} with the sampled TS.
}
\description{
In order to validate the model, I will select at least one TS sample randomly for each product category and store.
Also, because many TS are incomplete, I will use only those with less than 25\% of missing values.
}
