% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{split_dataset}
\alias{split_dataset}
\title{Split dataset into train and test sets}
\usage{
split_dataset(tsb, test_size)
}
\arguments{
\item{tsb}{A \code{tsibble} with the time series data.}

\item{test_size}{The proportion of data to use for testing.}
}
\value{
A \code{list} with train and test sets.
}
\description{
Split dataset into train and test sets
}
