% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_tsibble}
\alias{to_tsibble}
\title{Parse tibble to tsibble with defined index and keys}
\usage{
to_tsibble(tbl)
}
\arguments{
\item{tbl}{A \code{tibble}}
}
\value{
A \code{tsibble}
}
\description{
Parse tibble to tsibble with defined index and keys
}
