% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train_model}
\alias{train_model}
\title{Train Models}
\usage{
train_model(train, mods)
}
\arguments{
\item{train}{A \code{tsibble} with training data.}

\item{mods}{A named \code{list} of models to train.}
}
\value{
A \code{mable}.
}
\description{
Train Models
}
