% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{write_to_parquet}
\alias{write_to_parquet}
\title{Decorate function to add a message after saving data}
\usage{
write_to_parquet(data, path)
}
\arguments{
\item{data}{Data frame to be written to parquet.}

\item{path}{Path}
}
\description{
Decorate function to add a message after saving data
}
